Started working on feature boxes.
Ran into a new project concept that I was looking for; called natural language generation.

It follows the format will of my link-parser/grammar many projects outputs and the rule tagging will work perfect for article and news or cost per click ad generation based on financial document or other source of processing of a text on an identity or entity in SpclOps.

	new file:   NLP_Tools.pdf
	new file:   Possible/NLG/README.md
	new file:   feature-boxes.html